requirements.txt
click==8.1.7
pycurl==7.45.7
Flask==3.0.3
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pkce==1.0.3
python-jose[cryptography]==3.3.0
requests==2.32.3
Werkzeug==3.0.3
